requirements.txt
python-dotenv
watchdog
Flask
peewee
psycopg2-binary


python-dotenv==0.19.1
watchdog==2.1.3
Flask==2.0.1
peewee==3.17.1
psycopg2-binary==2.9.9
